requirements.txt
django
djangorestframework
markdown
django-filter
django-extensions
django-allauth
django-cors-headers
social-auth-app-django
rest-social-auth
dj-rest-auth

django-storages
drf-access-policy
drf-nested-routers
drf-yasg
pylint-django
django_prometheus
Celery
django-redis


django-rest-framework-social-oauth2
django-request-logging
django-health-check
django-oauth-toolkit
pytest-django
pytest-factoryboy
pytest-cov
python-decouple
django-q

django-storages
